move top-level surface array conversion func to header
This was previously local to the python bindings, preventing it from
being used in other C++ code.
